Change `tweets` type from array to string array in types.ts
In the types.ts file, the type of `tweets` field has been altered. Previously, it was an empty array, which was not providing satisfactory information about the data type it is intended to store. Now, it has been changed to a string array, clarifying that it is assumed to store strings.